requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
commonmark==0.9.1
Flask==1.1.2
hruid==0.0.3
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.3
jsonpickle==2.0.0
MarkupSafe==1.1.1
numpy==1.20.1
Pillow==8.1.2
Pygments==2.8.1
PyYAML==5.4.1
pyzmq==22.0.3
requests==2.25.1
rich==9.13.0
termcolor==1.1.0
torch==1.8.0
torchvision==0.9.0
typing-extensions==3.7.4.3
urllib3==1.26.4
Werkzeug==1.0.1
